@@ -2,18 +2,18 @@ name: xml_layout
 description: >
   XML layout for flutter.
   Layout your widgets with xml at runtime.
-version: 2.2.14
+version: 2.3.0
 #author:
 homepage: https://github.com/gsioteam/xml_layout
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  xml: '>=4.2.0'
-  analyzer: ">=0.40.6"
+  xml: ^4.5.0
+  analyzer: ^0.41.2
   dart_style: ^1.3.9
   build: ^1.5.0
 
