@@ -1,7 +1,7 @@
 
 class Collection {
   final Type collectionType;
-  final Type targetType;
+  final Type? targetType;
 
   const Collection(this.collectionType, [this.targetType]);
 }
